🎨  Improve application structure: New files created to simplify the structure. [skip actions]
