+++ app/notification/application_submission/dummy_data.json
@@ -112,4 +112,3 @@
         }
     ]
 }
-
